redirect al seguir y dejar de seguir, y comprobaciones

+++ FutbolScoutNet/sitio/views.py
@@ -131,18 +131,20 @@ def follow(request, username):
     current_user = request.user
     to_user = User.objects.get(username = username)
     to_user_id = to_user
-    rel = Relationship(from_user=current_user, to_user=to_user_id)
-    rel.save()
-    return redirect('home')
+    if (not Relationship.objects.filter(from_user = current_user.id, to_user=to_user_id).exists()):
+        rel = Relationship(from_user=current_user, to_user=to_user_id)
+        rel.save()
+    return redirect('profile', username=to_user.username)
 
 @login_required
 def unfollow(request, username):
     current_user = request.user
     to_user = User.objects.get(username = username)
     to_user_id = to_user
-    rel = Relationship.objects.filter(from_user = current_user.id, to_user=to_user_id).get()
-    rel.delete()
-    return redirect('home')
+    if (Relationship.objects.filter(from_user = current_user.id, to_user=to_user_id).exists()):
+        rel = Relationship.objects.filter(from_user = current_user.id, to_user=to_user_id).get()
+        rel.delete()
+    return redirect('profile', username=to_user.username)
 
 #EMIAL CONFIRMATION
